This patch removes the need for a content script at document start, which potentially delayed page load, and instead moved to using the extensions tab tracking API.
This actually calls us back earlier than the content script based notification.

We use this to sticky the browser action icon, and to re-synch the profiling
options to be consistent with the values set in the monitor UI. There may be a
small window of time where the profiling options are out of synch.

This is an interim solution. In practice the profiling options re-synching
happens before any script in the page runs (the resynch goes out before the
main page resource is even loaded), and was verified to run in time to catch 
inline script running in google.com.

R: zundel / http://gwt-code-reviews.appspot.com/130805/show