Fixes bug where stopping/starting monitoring causes the profiling state to be unsynched.
Starting and stopping will nuke the devtools client host. We need to resend the
profiling options as reflected in the UI.

This patch also allows log messages to show up in the trace tree. This is very
useful to know when things happen in the tree.

R: zundel / http://gwt-code-reviews.appspot.com/137802