This patch renames NetworkTimeLineModel to NetworkVisualizationModel.
This patch moves resource tracking in to the NetworkResourceModel which is
managed by DataModel. This allows us to be able to look up information about a
network resource from any Visualization.

This patch adds a color for the ResourceDataReceived event, fixes its
description, and connects the Details Table to the URL for the resource.

R: zundel
http://gwt-code-reviews.appspot.com/295801