The motivation behind this patch is that GPE needs some other information associated with each Jump-To-IDE request that it knows about when generating the symbol manifest.
This patch changes sourceViewServer to be a JavaScript object map instead of
just a URL String.

This map has a 'url' field corresponding to the source viewer server URL, and
any number of opaque values that we simply forward on as URL params when making
the request to the source viewer server.

R: zundel
http://gwt-code-reviews.appspot.com/456801/show